# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Trade Finder API
  version: "2.0"
  description: |
    OpenAPI specification for the AI Trade Finder project.

    Notes:
    - Authentication is not implemented (all endpoints are open).
    - Some controllers return empty bodies on error (e.g., 500 with no JSON).
    - Several fields are intentionally flexible objects (Map<String,Object>); those are modeled as `type: object` with `additionalProperties: true`.
servers:
  - url: http://localhost:8082
    description: Local default
...
paths:
  /api/v2/ai/analyze/trade:
    post:
      tags: [AI v2]
      summary: Structured trade analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeTradeRequest'
            examples:
              basic:
                value:
                  symbol: NQ
                  timeframe: 5m
                  marketData:
                    price: 21850
                    volume: 150000
                  indicators: [RSI, MACD, CISD]
      responses:
        '200':
          description: Trade signal (structured)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeSignalResponseDTO'
        '500':
          description: Internal error (often empty body)
  /api/v2/ai/analyze:
    post:
      tags: [AI v2]
      summary: Generic structured analysis (arbitrary JSON)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StructuredAnalysisRequest'
            examples:
              market_structure:
                value:
                  analysisType: MARKET_STRUCTURE
                  input: Analyze NQ price action from last 2 hours
                  systemInstructions: You are a market structure expert
                  maxTokens: 4000
                  temperature: 0.7
                  responseFormat: json_object
      responses:
        '200':
          description: Arbitrary JSON object
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '500':
          description: Internal error (often empty body)
  /api/v2/ai/chat/sessions:
    post:
      tags: [AI v2]
      summary: Create a chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatSessionRequest'
            examples:
              trade_discussion:
                value:
                  title: NQ Long Setup Discussion
                  type: TRADE_DISCUSSION
                  tradeId: 65a1b2c3d4e5f6g7h8i9j0k1
                  symbol: NQ
      responses:
        '201':
          description: Created chat session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '500':
          description: Internal error (often empty body)
  /api/v2/ai/chat/sessions/{sessionId}/messages:
    post:
      tags: [AI v2]
      summary: Send a message in a chat session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
            examples:
              question:
                value:
                  message: Why is this a high-confidence setup?
      responses:
        '200':
          description: AI reply
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '400':
          description: Session is not active (often empty body)
        '404':
          description: Not found (may be 500 depending on exception handling)
        '500':
          description: Internal error (often empty body)
  /api/v2/ai/chat/sessions/{sessionId}:
    get:
      tags: [AI v2]
      summary: Get chat session by ID
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Chat session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '404':
          description: Not found
  /api/v2/ai/chat/trades/{tradeId}/sessions:
    get:
      tags: [AI v2]
      summary: List chat sessions for a trade
      parameters:
        - in: path
          name: tradeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSession'
  /api/v2/ai/chat/sessions/{sessionId}/close:
    post:
      tags: [AI v2]
      summary: Close a chat session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Closed
        '500':
          description: Internal error (often empty body)
...
  /api/v2/ai/chat/discuss-trade:
    post:
      tags: [AI v2]
      summary: Create a trade discussion session (optionally send initial question)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscussTradeRequest'
            examples:
              start:
                value:
                  tradeId: 65a1b2c3d4e5f6g7h8i9j0k1
                  initialQuestion: What are the risks with this setup?
      responses:
        '201':
          description: Created chat session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '500':
          description: Internal error (often empty body)
...
components:
  schemas:
...
    TradeSignalResponseDTO:
      type: object
      properties:
        status:
          type: string
          enum: [TRADE_IDENTIFIED, NO_SETUP, INSUFFICIENT_DATA, ERROR]
        direction:
          type: string
          enum: [LONG, SHORT]
        symbol: { type: string }
        timeframe: { type: string }
        confidence:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
        entry:
          $ref: '#/components/schemas/TradeSignalEntryInfo'
        stop:
          $ref: '#/components/schemas/TradeSignalStopInfo'
        targets:
          type: array
          items:
            $ref: '#/components/schemas/TradeSignalTargetInfo'
        risk_reward: { type: string, nullable: true }
        narrative: { type: string, nullable: true }
        trigger_conditions:
          type: array
          items: { type: string }
          nullable: true
        invalidations:
          type: array
          items: { type: string }
          nullable: true
        session_label: { type: string, nullable: true }
        analysis_timestamp: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    TradeSignalEntryInfo:
      type: object
      properties:
        zone_type:
          type: string
          enum: [FVG_CE, IFVG, OB, BREAKER, MITIGATION]
        zone: { type: string, nullable: true }
        price: { type: number, format: double, nullable: true }
        method: { type: string, nullable: true }
    TradeSignalStopInfo:
      type: object
      properties:
        placement: { type: string, nullable: true }
        price: { type: number, format: double, nullable: true }
        reasoning: { type: string, nullable: true }
    TradeSignalTargetInfo:
      type: object
      properties:
        level: { type: string, nullable: true }
        price: { type: number, format: double, nullable: true }
        description: { type: string, nullable: true }
    AnalyzeTradeRequest:
      type: object
      required: [symbol]
      properties:
        symbol: { type: string }
        timeframe: { type: string, nullable: true }
        marketData:
          type: object
          nullable: true
          additionalProperties: true
        indicators:
          type: array
          nullable: true
          items: { type: string }
    StructuredAnalysisRequest:
      type: object
      required: [analysisType, input]
      properties:
        analysisType: { type: string }
        input: { type: string }
        systemInstructions: { type: string, nullable: true }
        maxTokens: { type: integer, format: int32, nullable: true }
        temperature: { type: number, format: double, nullable: true }
        responseFormat:
          type: string
          nullable: true
          description: "Present in DTO; controller currently parses output as JSON regardless."
    CreateChatSessionRequest:
      type: object
      properties:
        title: { type: string, nullable: true }
        type:
          type: string
          description: Recommended values TRADE_DISCUSSION, MARKET_ANALYSIS, GENERAL
        tradeId: { type: string, nullable: true }
        symbol: { type: string, nullable: true }
    ChatMessageRequest:
      type: object
      required: [message]
      properties:
        message: { type: string }
    ChatMessageResponse:
      type: object
      properties:
        sessionId: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        model: { type: string, nullable: true }
    DiscussTradeRequest:
      type: object
      required: [tradeId]
      properties:
        tradeId: { type: string }
        initialQuestion: { type: string, nullable: true }
    ChatSession:
      type: object
      properties:
        id: { type: string }
        title: { type: string, nullable: true }
        type: { type: string }
        tradeId: { type: string, nullable: true }
        symbol: { type: string, nullable: true }
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        context: { type: string, nullable: true }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
        createdBy: { type: string, nullable: true }
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          description: user | assistant
        content: { type: string }
        timestamp: { type: string, format: date-time }
        tokens: { type: integer, format: int32, nullable: true }
        model: { type: string, nullable: true }
...